Revert not well understood potential fix
Passing the user returned from self.login_user() to self.get_next_url()
can perhaps impact the value returned by self.get_next_url in a way that
fixes an issue. This is not obvious to me if it will or not, it just
seems likely that it makes sense to do.

Without further information and experimentation, I'll revert doing that
as it could lead to a breaking change for users.