fix: xsrf error on JupyterHub 4.0.0
JupyterHub 4.0.0 extended the verification of the _xsrf parameter in POST
requests. But LTI requests do not include this parameter, and requests
triggered the following error:

    403: Forbidden '_xsrf' argument missing from POST

LTIAuthenticator thus was not compatible with JupyterHub 4.0.0. This PR
resolves the issue for LTI 1.1 by ignoring the xsrf check; a similar
change should be introduced for the LTI 1.3 authenticator.

See the tornado docs for more information on `check_xsrf_cookie`:
https://www.tornadoweb.org/en/stable/web.html#tornado.web.RequestHandler.check_xsrf_cookie

Close #157.